pyproject.toml
[project]
name = "apng2gif"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "mypy==1.16.1",
    "pillow==11.2.1",
    "pyinstaller==6.14.1",
]
